Use member groups to simplify documentation
For lists of fields that are copied or derived from the video bitstreams, use
Doxygen member groups to document them once as a block, rather than copying the
text "Copy of the <whatever> bitstream field." all over the place.  This groups
the fields together in the HTML.

Reviewed-by: Christian König <christian.koenig@amd.com>
v2: Rebase on top of José's HEVC work.
Signed-off-by: Aaron Plattner <aplattner@nvidia.com>
Reviewed-by: José Hiram Soltren <jsoltren@nvidia.com>