Change HEVC profile numbers to 100 through 104
libvdpau 1.0 contained an error in its HEVC picture info structures.  Rather
than try to maintain backward compatibility with the incorrect definition, the
existing VdpPictureInfoHEVC was updated to contain the fixed definition.  Since
the new structure is no longer compatible with the ABI defined by libvdpau 1.0,
change the profile numbers for HEVC so that software built against the incorrect
definition will not recognize the new profiles.

Signed-off-by: Aaron Plattner <aplattner@nvidia.com>
Reviewed-by: José Hiram Soltren <jsoltren@nvidia.com>