Add HEVC 444 support in VDPAU API
Added new VdpPictureInfoHEVC444 structure for HEVC444 support.
having the SPS,PPS range extension variables that are defined for HEVC
444. VdpPictureInfoHEVC is part of VdpPictureInfoHEVC444.

New VdpYCbCr Formats are added to be used in get/putbits for YUV 4:4:4
surfaces.

Added the capability bits for chromatypes supported.

Add support to return the supported chroma types in
VdpDecoderQueryProfileCapability API. The supported chroma types
are returned in a bitmask.

Signed-off-by: Aaron Plattner <aplattner@nvidia.com>