meson: Remove unnecessary 'install' parameter from configure_file()
The 'install' parameter to the configure_file() function explicitly controls
whether the file is installed, but omitting it just infers whether to install
the file based on the presence or absence of the 'install_dir' parameter. This
parameter requires Meson 0.50 or newer:

 WARNING: Project specifies a minimum meson_version '>=0.41' but uses features which were added in newer versions:
  * 0.50.0: {'install arg in configure_file'}

We don't need to specify 'install' for this particular file because
'install_dir' is not set, so just remove it to drop the Meson requirement back
down to 0.41.

Signed-off-by: Aaron Plattner <aplattner@nvidia.com>